Replace likelihood_field package with cartesian_likelihood_field package and rename it with likelihood_grid

+++ likelihood_grid/include/polarcord.h
@@ -6,21 +6,21 @@
 class PolarPose
 {
 public:
-    float range;
-    float angle;
+    double range;
+    double angle;
     PolarPose():range(0.0), angle(0.0){;}
-    PolarPose(const float r, const float a ):range(r), angle(a){;}
-    inline void fromCart(const float x, const float y)
+    PolarPose(const double r, const double a ):range(r), angle(a){;}
+    inline void fromCart(const double x, const double y)
     {
         range = sqrt((x * x) + (y * y));
         angle = atan2(y, x);
     }
-    inline void toCart(float &x, float &y) {
+    inline void toCart(double &x, double &y) {
         x = range * cos(angle);
         y = range * sin(angle);
     }
 
-    inline double distance(const float r, const float a)
+    inline double distance(const double r, const double a)
     {
         return sqrt( pow(range,2) + pow(r,2) - 2*range*r*cos(angle - a));
     }
@@ -28,7 +28,7 @@ public:
 
 inline static double toRadian(double t)
 {
-    return t*M_PI/180;
+    return t*M_PI/180.0;
 }
 
 #endif
